requirements.txt
email-validator>=2.2.0
flask-migrate>=4.1.0 
flask>=3.1.1 
flask-sqlalchemy>=3.1.1 
google-generativeai>=0.8.5 
pdf2image>=1.17.0 
pillow>=11.2.1 
pypdf2>=3.0.1 
pytesseract>=0.3.13 
sqlalchemy>=2.0.41 
werkzeug>=3.1.3 
requests>=2.32.3 
python-dotenv==1.1.0